+++ scripts/locale-converter.py
@@ -28,10 +28,10 @@ def convert(source, destination, current = None, level = 1):
         path = os.path.join(dir, name)
         
         if os.path.isfile(path):
-            if path.endswith('profilewizard.dtd'):
-                messages.extend(convert_dtd(path, dir))
-            #if path.endswith('.properties'):
-             #   messages.extend(convert_prop(path, dir))
+            #if path.endswith('profilewizard.dtd'):
+             #   messages.extend(convert_dtd(path, dir))
+            if path.endswith('autobackup.properties'):
+                messages.extend(convert_prop(path, dir))
 
         else:
             messages.extend(convert(source, destination, path, level+1))